+++ src/models.rs
@@ -280,8 +280,8 @@ pub_struct!(Debug, Serialize, Deserialize; BoostTable {
     img_url: String,
     name: String,
     hidden: bool,
-    num_of_winners: i64,
-    token_decimals: i64,
+    num_of_winners: i32,
+    token_decimals: i32,
 });
 
 pub_struct!(Debug, Serialize, Deserialize; NftBalance {
@@ -369,8 +369,8 @@ pub_struct!(Deserialize; CreateBoostQuery {
     quest_id: i32,
     amount: i32,
     token: String,
-    num_of_winners: i64,
-    token_decimals: i64,
+    num_of_winners: i32,
+    token_decimals: i32,
     name: String,
     img_url: String,
     expiry: i64,
